fix: Routes and splash screen rendering
