+++ app/(auth)/register.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { ImageBackground, Text, TextInput, TouchableOpacity, View } from 'react-native';
+import { ImageBackground, ScrollView, Text, TextInput, TouchableOpacity, View } from 'react-native';
 import { KeyRound, Mail, User } from 'lucide-react-native';
 import { COLORS } from '@/constants/Colors';
 import Animated, { FadeInUp } from 'react-native-reanimated';
@@ -9,6 +9,8 @@ import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import axios from 'axios';
 import { useRouter } from 'expo-router';
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import type { response } from '@/types/authResponse';
 
 const schema = z.object({
   name: z.string({required_error: "Insira um nome"}).min(2, { message: "O nome precisa de pelo menos 2 caracteres" }),
@@ -27,17 +29,17 @@ export default function Register() {
 
   const [errorMessage, setErrorMessage] = useState<string>("")
 
-  
-
   const onSubmit = async (data: z.infer<typeof schema>) => {
     await axios.post(`${process.env.EXPO_PUBLIC_API_URL}/register`, {
       name: data.name,
       email: data.email,
       password: data.password
     })    
-    .then((res) => {
+    .then((res: response) => {
       console.log('Success response:', res.data);
-      route.push('/home')
+      route.replace('/(home)')
+      AsyncStorage.setItem("token", res.data.token);
+      AsyncStorage.setItem("user", JSON.stringify(res.data.user));
     })
     .catch((error) => {
       if (error.response) {
@@ -61,12 +63,12 @@ export default function Register() {
   },[onSubmit])
 
   return (
-    <View className='flex h-screen bg-off-black'>
+    <View className='flex h-full bg-off-black pb-4'>
       <View className='flex h-1/2'>
         <ImageBackground
           source={require("@/assets/images/stadium_gradient_bottom-transformed.png")}
           className='h-full w-full'
-          resizeMode='contain'
+          resizeMode='cover'
         >
           <Animated.View entering={FadeInUp.duration(1000)} className='h-full flex items-center justify-end gap-2'>
             <Text className='text-off-white text-4xl font-thin'>Já está criando</Text>
@@ -74,7 +76,7 @@ export default function Register() {
           </Animated.View>
         </ImageBackground>
       </View>
-
+      <ScrollView>
       <View className='flex h-full bg-off-black items-center pt-5'>
         <Text className='flex self-start text-off-white text-2xl pl-5 mb-2'>Cadastrar</Text>
 
@@ -90,6 +92,7 @@ export default function Register() {
                 placeholder="Nome"
                 value={value}
                 onChangeText={onChange}
+                autoCapitalize='words'
               />
             )}
           />
@@ -108,6 +111,8 @@ export default function Register() {
                 placeholder="Email"
                 value={value}
                 onChangeText={onChange}
+                autoCapitalize='none'
+                autoComplete='email'
               />
             )}
           />
@@ -127,6 +132,7 @@ export default function Register() {
                 secureTextEntry
                 value={value}
                 onChangeText={onChange}
+                autoCapitalize='none'
               />
             )}
           />
@@ -141,6 +147,7 @@ export default function Register() {
           <Text className='text-off-white mt-1'>Já possui conta? <Link className='text-green' href="/">Entrar</Link></Text>
         </View>
       </View>
+      </ScrollView>
     </View>
   );
 }
