+++ app/(home)/_layout.tsx
@@ -1,0 +1,17 @@
+import { Tabs } from 'expo-router';
+import { Plus } from 'lucide-react-native';
+
+export default function HomeLayout() {
+  return (    
+    <Tabs >
+      <Tabs.Screen 
+        name="index"
+        options={{
+          headerShown: false, 
+          tabBarIcon: (() => <Plus />),
+          title: "Home",
+        }} 
+      />
+    </Tabs>
+  );
+}
