--- app/home/index.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Text, View } from 'react-native';
-import { SafeAreaView } from 'react-native-safe-area-context';
-
-const Home = () => {
-  return(
-    <SafeAreaView>
-      <Text>Hello world</Text>
-    </SafeAreaView>
-  );
-}
-
-export default Home;
